JavaScript (v3): EventBridge - Rename files.
