+++ javascriptv3/example_code/eventbridge/src/libs/client.js
