+++ javascriptv3/example_code/eventbridge/src/put-rule.js
@@ -5,35 +5,32 @@ ABOUT THIS NODE.JS EXAMPLE: This example works with the AWS SDK for JavaScript v
 which is available at https://github.com/aws/aws-sdk-js-v3.
 
 Purpose:
-putTargets.js demonstrates how to add or update a target to an Amazon EventBridge rule.
+putRule.js demonstrates how to create or update an Amazon EventBridge rule.
 
 Inputs (replace in code):
-- LAMBDA_FUNCTION_ARN
+- IAM_ROLE_ARN
 
 Running the code:
-node putTargets.js
+node putRule.js
 */
-// snippet-start:[eventBridge.JavaScript.eb.putTargetsV3]
+// snippet-start:[eventBridge.JavaScript.eb.putRuleV3]
 
 // Import required AWS SDK clients and commands for Node.js.
-import { PutTargetsCommand } from "@aws-sdk/client-eventbridge";
-import { ebClient } from "./libs/eventBridgeClient.js";
+import { PutRuleCommand } from "@aws-sdk/client-eventbridge";
+import { ebClient } from "./libs/client.js";
 
 // Set the parameters.
 export const params = {
-  Rule: "DEMO_EVENT",
-  Targets: [
-    {
-      Arn: "LAMBDA_FUNCTION_ARN", //LAMBDA_FUNCTION_ARN
-      Id: "myCloudWatchEventsTarget",
-    },
-  ],
+  Name: "DEMO_EVENT",
+  RoleArn: "IAM_ROLE_ARN", //IAM_ROLE_ARN
+  ScheduleExpression: "rate(5 minutes)",
+  State: "ENABLED",
 };
 
 export const run = async () => {
   try {
-    const data = await ebClient.send(new PutTargetsCommand(params));
-    console.log("Success, target added; requestID: ", data);
+    const data = await ebClient.send(new PutRuleCommand(params));
+    console.log("Success, scheduled rule created; Rule ARN:", data);
     return data; // For unit tests.
   } catch (err) {
     console.log("Error", err);
@@ -41,6 +38,6 @@ export const run = async () => {
 };
 // Uncomment this line to run execution within this file.
 // run();
-// snippet-end:[eventBridge.JavaScript.eb.putTargetsV3]
+// snippet-end:[eventBridge.JavaScript.eb.putRuleV3]
 // For unit tests only.
 // module.exports ={run, params};
