+++ javascriptv3/example_code/eventbridge/tests/babel.config.js
@@ -1,3 +1,4 @@
 module.exports = {
   presets: ["@babel/preset-env"],
 };
+
