+++ javascriptv3/example_code/eventbridge/tests/jest.config.js
@@ -6,3 +6,4 @@ module.exports = {
     "^.+\\.(js|jsx)$": "babel-jest",
   },
 };
+
