--- javascriptv3/example_code/eventbridge/tests/package.json
@@ -1,20 +1,0 @@
-{
-  "name": "aws-sdk-v3-cloudwatch-events-examples",
-  "version": "1.0.0",
-  "main": "index.js",
-  "repository": "git@github.com/awsdocs/aws-doc-sdk-examples/tree/master/javascriptv3/example_code/cloudwatch-events.git",
-  "author": "Brian Murray <brmur@amazon.com>, Alex Forsyth <alex-git@amazon.com>",
-  "license": "Apache 2.0",
-  "dependencies": {
-    "@aws-sdk/client-eventbridge": "^3.387.0"
-  },
-  "devDependencies": {
-    "@babel/preset-env": "^7.14.4",
-    "@types/node": "^14.0.23",
-    "jest": "^26.6.3",
-    "regenerator-runtime": "^0.13.7"
-  },
-  "scripts": {
-    "test": "jest"
-  }
-}
